+++ packages/fleather/test/testing.dart
@@ -4,7 +4,6 @@ import 'package:fleather/src/widgets/text_selection.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:parchment_delta/parchment_delta.dart';
 
 var delta = Delta()..insert('This House Is A Circus\n');
 
@@ -220,17 +219,20 @@ class TestUpdateWidgetState extends State<TestUpdateWidget> {
             child: const Text('Change state'),
           ),
           if (widget.toolbarBuilder != null) widget.toolbarBuilder!(context),
-          widget.testField
-              ? FleatherField(
-                  controller: widget.controller ??
-                      FleatherController(document: widget.document),
-                  focusNode: focusNode,
-                )
-              : FleatherEditor(
-                  controller: widget.controller ??
-                      FleatherController(document: widget.document),
-                  focusNode: focusNode,
-                ),
+          Expanded(
+            child: widget.testField
+                ? FleatherField(
+                    controller: widget.controller ??
+                        FleatherController(document: widget.document),
+                    focusNode: focusNode,
+                  )
+                : FleatherEditor(
+                    controller: widget.controller ??
+                        FleatherController(document: widget.document),
+                    focusNode: focusNode,
+                    scrollable: true,
+                  ),
+          ),
         ],
       );
 }