+++ packages/fleather/test/widgets/editor_toolbar_test.dart
@@ -1,7 +1,6 @@
 import 'package:fleather/fleather.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:parchment_delta/parchment_delta.dart';
 
 import '../testing.dart';
 
@@ -50,13 +49,15 @@ Widget widget(FleatherController controller, {bool withBasic = false}) {
           )
         ],
       );
+  final editorKey = GlobalKey<EditorState>();
   return MaterialApp(
     home: Material(
       child: Column(children: [
         if (withBasic)
-          FleatherToolbar.basic(controller: controller)
+          FleatherToolbar.basic(controller: controller, editorKey: editorKey)
         else
           FleatherToolbar(
+            editorKey: editorKey,
             children: [
               ToggleStyleButton(
                 attribute: ParchmentAttribute.bold,
@@ -110,6 +111,7 @@ Widget widget(FleatherController controller, {bool withBasic = false}) {
             child: FleatherEditor(
           controller: controller,
           maxContentWidth: 800,
+          editorKey: editorKey,
         ))
       ]),
     ),
@@ -148,6 +150,10 @@ void main() {
 
       expect(tester.widget<FLIconButton>(rawRedoButton).onPressed, isNull);
       expect(controller.document.toDelta(), Delta()..insert('Hello world\n'));
+
+      final editor = tester.state<RawEditorState>(find.byType(RawEditor));
+      expect(editor.effectiveFocusNode.hasFocus, false);
+      expect(editor.hasConnection, false);
     });
 
     testWidgets('Bold', (tester) async {
@@ -204,6 +210,10 @@ void main() {
 
       expect(controller.document.toDelta().last,
           Operation.insert('\n', {'heading': 6}));
+
+      final editor = tester.state<RawEditorState>(find.byType(RawEditor));
+      expect(editor.effectiveFocusNode.hasFocus, true);
+      expect(editor.hasConnection, true);
     });
 
     testWidgets('Indentation', (tester) async {
@@ -227,6 +237,10 @@ void main() {
       await tester.pumpAndSettle(throttleDuration);
       expect(controller.document.toDelta().last,
           Operation.insert('Hello world\n'));
+
+      final editor = tester.state<RawEditorState>(find.byType(RawEditor));
+      expect(editor.effectiveFocusNode.hasFocus, true);
+      expect(editor.hasConnection, true);
     });
 
     testWidgets('Link', (tester) async {
@@ -260,6 +274,10 @@ void main() {
           controller.document.toDelta().first,
           Operation.insert(
               'Hello', {'a': 'https://fleather-editor.github.io'}));
+
+      final editor = tester.state<RawEditorState>(find.byType(RawEditor));
+      expect(editor.effectiveFocusNode.hasFocus, true);
+      expect(editor.hasConnection, true);
     });
 
     testWidgets('Horizontal rule', (tester) async {
@@ -278,6 +296,10 @@ void main() {
 
       expect(controller.document.toDelta().elementAt(1),
           Operation.insert({'_type': 'hr', '_inline': false}));
+
+      final editor = tester.state<RawEditorState>(find.byType(RawEditor));
+      expect(editor.effectiveFocusNode.hasFocus, true);
+      expect(editor.hasConnection, true);
     });
 
     testWidgets('Basic toolbar', (tester) async {
@@ -334,6 +356,10 @@ void main() {
       await tester.pumpAndSettle(throttleDuration);
       expect(controller.document.toDelta().first,
           Operation.insert('Hello', {'bg': Colors.black.value}));
+
+      final editor = tester.state<RawEditorState>(find.byType(RawEditor));
+      expect(editor.effectiveFocusNode.hasFocus, true);
+      expect(editor.hasConnection, true);
     });
 
     testWidgets('Text color', (tester) async {
@@ -365,6 +391,10 @@ void main() {
       await tester.pumpAndSettle(throttleDuration);
       expect(controller.document.toDelta().first,
           Operation.insert('Hello', {'fg': Colors.black.value}));
+
+      final editor = tester.state<RawEditorState>(find.byType(RawEditor));
+      expect(editor.effectiveFocusNode.hasFocus, true);
+      expect(editor.hasConnection, true);
     });
 
     testWidgets('updating editor toolbar remove overlay entry if any',
@@ -493,6 +523,10 @@ Future<void> performToggle(WidgetTester tester, FleatherController controller,
   await tester.pumpAndSettle(throttleDuration);
   expect(
       controller.document.toDelta().first, Operation.insert('Hello world\n'));
+
+  final editor = tester.state<RawEditorState>(find.byType(RawEditor));
+  expect(editor.effectiveFocusNode.hasFocus, true);
+  expect(editor.hasConnection, true);
 }
 
 void onErrorIgnoreOverflowErrors(